demo/backend/requirements.txt
anyio==4.0.0
audioread==3.0.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
decorator==5.1.1
dnspython==2.4.2
ecdsa==0.18.0
exceptiongroup==1.1.3
faiss-cpu==1.7.4
fastapi==0.95.1
filelock==3.9.0
h11==0.14.0
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
librosa==0.9.0
llvmlite==0.40.1
MarkupSafe==2.1.2
motor==3.3.1
mpmath==1.2.1
networkx==3.0
numba==0.57.1
numpy==1.23.5
packaging==23.1
passlib==1.7.4
Pillow==9.3.0
platformdirs==3.10.0
pooch==1.7.0
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.12
pydub==0.25.1
pymongo==4.5.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
resampy==0.4.2
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
starlette==0.26.1
sympy==1.11.1
threadpoolctl==3.2.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.22.0
